refactor: split file
The logic for resolving dependencies was in the npm-client file. The file was getting pretty large because of this and also mixed levels of abstraction.

Moved dependency-resolving logic to own file
